finished most of feeds and feed replis

+++ routes/classroom.js
@@ -8,7 +8,6 @@ const auth = require("../middleware/auth");
 // models
 const Classroom = require("../models/Classroom");
 const Department = require("../models/Department");
-const Staff = require("../models/Staff");
 
 const router = Router({ mergeParams: true }); //initialize the route
 
@@ -17,7 +16,7 @@ const router = Router({ mergeParams: true }); //initialize the route
  *  @method GET
  *  ?route -->/departments/:id/classrooms
  *  @param none
- *  @access private
+ *  @access detpHead
  */
 
 router.get("/", deptHead, async (req, res) => {
@@ -37,7 +36,7 @@ router.get("/", deptHead, async (req, res) => {
  *  @method POST
  *  ?route --> /departments/:id/classrooms
  *  @param {name: <classroom name>, deptId: <department under which the class is being created>}
- *  @access private
+ *  @access deptHead
  */
 
 router.post("/", deptHead, async (req, res) => {
@@ -64,7 +63,7 @@ router.post("/", deptHead, async (req, res) => {
  *  @method GET
  *  ?route --> /departments/:id/classrooms/:cid
  *  @param none
- *  @access private
+ *  @access auth
  */
 
 router.get("/:cid", auth, async (req, res) => {
@@ -74,6 +73,10 @@ router.get("/:cid", auth, async (req, res) => {
       .populate({
         path: "homeRoomTeacher department studentMembers staffMembers",
         select: "eduMail name",
+      })
+      .populate({
+        path: "feeds",
+        select: "title body",
       })
       .exec();
     res.json({ classroom });
@@ -88,7 +91,7 @@ router.get("/:cid", auth, async (req, res) => {
  *  @method PUT
  *  ?route --> /departments/:id/classrooms/:cid/staffs
  *  @param {staffs: <array of staffs id>}
- *  @access private
+ *  @access deptHead
  */
 
 router.put("/:cid/staffs", deptHead, async (req, res) => {
@@ -112,7 +115,7 @@ router.put("/:cid/staffs", deptHead, async (req, res) => {
  *  @method PUT
  *  ?route --> /departments/:id/classrooms/:cid/students
  *  @param {students: <array of students id>}
- *  @access private
+ *  @access classCharge
  */
 
 router.put("/:cid/students", classCharge, async (req, res) => {
@@ -136,7 +139,7 @@ router.put("/:cid/students", classCharge, async (req, res) => {
  *  @method PUT
  *  ?route --> /departments/:id/classrooms/:cid/homeroom
  *  @param {staffId: <id of staff>}
- *  @access private
+ *  @access deptHead
  */
 
 router.put("/:cid/homeroom", deptHead, async (req, res) => {
@@ -163,7 +166,7 @@ router.put("/:cid/homeroom", deptHead, async (req, res) => {
  *  @method DELETE
  *  ?route --> /departments/:id/classrooms/:cid/
  *  @param none
- *  @access private
+ *  @access deptHead
  */
 
 router.delete("/:cid", deptHead, async (req, res) => {
@@ -171,6 +174,7 @@ router.delete("/:cid", deptHead, async (req, res) => {
 
   try {
     const department = await Department.findById(id);
+    if (!department) return res.sendStatus(404);
     const classIndex = department.classrooms.indexOf(cid);
     await Classroom.findByIdAndDelete(cid);
     department.classrooms.splice(classIndex, 1);
